23w07a

+++ minecraft/src/net/minecraft/client/gui/widget/SimplePositioningWidget.java
@@ -103,7 +103,7 @@ public class SimplePositioningWidget extends WrapperWidget {
 	}
 
 	public static void setPos(int low, int high, int length, Consumer<Integer> setter, float relative) {
-		int i = MathHelper.lerp(relative, 0, high - length);
+		int i = (int)MathHelper.lerp(relative, 0.0F, (float)(high - length));
 		setter.accept(low + i);
 	}
